files moved to echo folder
