+++ reactapp/src/ECHO/ProfessorsAvailability/index.js
@@ -1,0 +1,3 @@
+import ProfessorsAvailability from './ProfessorsAvailability';
+
+export default ProfessorsAvailability;
