+++ reactapp/src/ECHO/pocetni.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-import "./App.css";
+import "../App.css";
 import Dani from "./dani.js";
 
 class Pocetni extends Component {
